+++ NEWS.md
@@ -1,7 +1,8 @@
 # rasterpic (development version)
 
-- `rasterpic_img()` is now an S3 generic, with methods for the supported input
-  classes.
+- `rasterpic_img()` is now an S3 generic, with methods for the supported
+  input classes.
+- `rasterpic_img()` gains support for `stars` objects.
 
 # rasterpic 0.4.0